feat(mobile): add utility for date and time formatting
